# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LuraDB API",
    "description": "REST-native multi-model database — KeyValue- und JSON-Engine. `version` ist die API-Contract-Version (siehe API_VERSION), unabhängig von der Server-Version; letztere steht in der Extension `x-luradb-server-version`. Kompatibilitäts-Ranges: api/COMPATIBILITY.md. Laufzeit-Check: GET /version.",
    "contact": {
      "name": "Heiko Wein"
    },
    "license": {
      "name": "FSL-1.1-ALv2"
    },
    "version": "0.1.0",
    "x-luradb-server-version": "0.1.0"
  },
  "paths": {
...
    "/store-api/auth/users": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Gibt alle angelegten User zurück (Admins und reguläre User).",
        "description": "API Keys werden nicht ausgegeben — nur Name, Rolle und Erstellungszeitpunkt.",
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "Liste aller User (ohne API Keys)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Legt einen neuen User mit der Rolle `User` an.",
        "description": "Der API Key wird **ausschließlich in dieser Response** zurückgegeben und danach nicht mehr gespeichert.\nNur Admins dürfen diesen Endpunkt aufrufen.",
        "operationId": "create_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User angelegt. API Key einmalig in der Response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ungültiger Name"
          },
          "409": {
            "description": "User existiert bereits"
          }
        }
      }
    },
    "/store-api/auth/users/{name}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Löscht einen User und alle seine Domain-Permissions.",
        "description": "Der API Key des Users wird sofort ungültig — laufende Requests mit dem alten Key erhalten danach `401`.",
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User und alle Permissions gelöscht"
          },
          "404": {
            "description": "User nicht gefunden"
          }
        }
      }
    },
    "/store-api/auth/users/{name}/permissions": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Setzt oder überschreibt die Zugriffsberechtigung eines Users auf eine Domain.",
        "description": "`access` muss `\"read\"`, `\"write\"` oder `\"ddl\"` sein — jede Stufe schließt die\nniedrigeren ein. Für `kv` muss die Domain existieren; `json`/`rel` prüfen nur den Namen.",
        "operationId": "set_permission",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Permission gesetzt"
          },
          "400": {
            "description": "Ungültige access- oder domain-Angabe"
          },
          "404": {
            "description": "User oder Domain nicht gefunden"
          }
        }
      }
    },
    "/store-api/auth/users/{name}/permissions/{domain}": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Entzieht einem User die Zugriffsberechtigung auf eine bestimmte Domain.",
        "description": "`?store_type=json`/`rel` entzieht eine JSON-/rel-Domain-Permission (Default: kv).\nNach diesem Aufruf erhalten Requests des Users auf diese Domain `403 Forbidden`.",
        "operationId": "remove_permission",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "domain",
            "in": "path",
            "description": "Domain-Name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "store_type",
            "in": "query",
            "description": "'kv' (Default), 'json' oder 'rel'",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Permission entzogen"
          },
          "404": {
            "description": "Permission nicht gefunden"
          }
        }
      }
    },
    "/store-api/auth/users/{name}/rotate-key": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Generiert einen neuen API Key für den User und invalidiert den alten sofort.",
        "description": "Der neue Key wird **ausschließlich in dieser Response** zurückgegeben.\nVerwende dies bei Key-Leaks oder regelmäßiger Key-Rotation.",
        "operationId": "rotate_key",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Neuer API Key generiert (einmalig sichtbar)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RotateKeyResponse"
                }
              }
            }
          },
          "404": {
            "description": "User nicht gefunden"
          }
        }
      }
    },
...
    "/store-api/domains/{name}": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Returns metadata for a single domain by name. Returns 404 if it does not exist.",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Domains"
        ],
        "summary": "Schedules a domain and all its data for deletion.",
        "description": "The domain becomes immediately inaccessible; the actual purge runs in the background.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge started"
          },
          "404": {
            "description": "Domain not found"
          }
        }
      }
    },
...
    "/store-api/json/domains/{name}": {
      "get": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Returns metadata for a single JSON domain, including its document count.",
        "description": "Domains in `deleting` state answer with 410 Gone.",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "JSON domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonDomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is being deleted (background purge running)"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      },
      "delete": {
        "tags": [
          "JSON Document Store"
        ],
        "summary": "Marks a JSON domain as deleting. Documents become inaccessible immediately.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "JSON domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge follows"
          },
          "404": {
            "description": "Domain not found"
          },
          "503": {
            "description": "JSON engine disabled"
          }
        }
      }
    },
...
    "/store-api/rel/domains/{name}": {
      "get": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Returns metadata for a single relational domain. A domain in `deleting`",
        "description": "state answers 410 Gone (rel/013 finishes the physical purge).",
        "operationId": "get_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Relational domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelDomainResponse"
                }
              }
            }
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is being deleted (background purge running)"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      },
      "delete": {
        "tags": [
          "Relational Domains"
        ],
        "summary": "Marks a relational domain as deleting. Its tables/views become",
        "description": "inaccessible immediately; physical cleanup follows in rel/013.",
        "operationId": "delete_domain",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Relational domain name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Deletion accepted — background purge follows (rel/013)"
          },
          "404": {
            "description": "Domain not found"
          },
          "410": {
            "description": "Domain is already being deleted"
          },
          "503": {
            "description": "Relational engine disabled"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Username (1-50 chars, only [a-zA-Z0-9_-])."
          }
        }
      },
      "CreateUserResponse": {
        "type": "object",
        "required": [
          "name",
          "role",
          "api_key"
        ],
        "properties": {
          "api_key": {
            "type": "string",
            "description": "API Key (nur einmalig sichtbar — danach nicht mehr abrufbar)."
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "DomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "User-visible domain name."
          }
        }
      },
...
      "JsonDomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "document_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of documents — only set on the detail endpoint.",
            "nullable": true,
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Lifecycle state: \"active\" or \"deleting\" (background purge running)."
          }
        }
      },
...
      "RelDomainResponse": {
        "type": "object",
        "required": [
          "name",
          "created_at",
          "state"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Creation timestamp (Unix seconds).",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "description": "Lifecycle state: \"active\" or \"deleting\" (background purge running, rel/013).\nAlways present — so rel/013's later `list_domains` visibility change\n(surfacing `Deleting` domains too) needs no REST-contract change."
          }
        }
      },
      "RotateKeyResponse": {
        "type": "object",
        "required": [
          "name",
          "api_key"
        ],
        "properties": {
          "api_key": {
            "type": "string",
            "description": "Neuer API Key (nur einmalig sichtbar — danach nicht mehr abrufbar)."
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "SetPermissionRequest": {
        "type": "object",
        "required": [
          "domain",
          "access"
        ],
        "properties": {
          "access": {
            "type": "string",
            "description": "Zugriffsebene: `\"read\"`, `\"write\"` oder `\"ddl\"` (spec rel/011)."
          },
          "domain": {
            "type": "string",
            "description": "Name der Domain."
          },
          "store_type": {
            "type": "string",
            "description": "Store-Typ der Domain: `\"kv\"` (Default), `\"json\"` (spec json/012) oder `\"rel\"` (spec rel/011).",
            "nullable": true
          }
        }
      },
...
      "UserListItem": {
        "type": "object",
        "required": [
          "name",
          "role",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key wie in `luradb.toml` konfiguriert (Admins) oder wie bei User-Anlage / Key-Rotation erhalten. Nur bei `auth.enabled = true` erforderlich."
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}